fix a bunch of bugs, refactor
