+++ src/main/java/org/quiltmc/enigma_plugin/proposal/FallbackProposalService.java
@@ -9,7 +9,7 @@ import org.quiltmc.enigma_plugin.index.JarIndexer;
 public class FallbackProposalService extends NameProposerService {
 	public FallbackProposalService(JarIndexer indexer, EnigmaServiceContext<NameProposalService> context) {
 		super();
-		this.addIfEnabled(context, Arguments.DISABLE_MOJMAP, () -> new MojmapNameProposer(context.getSingleArgument(Arguments.MOJMAP_PATH), context.getSingleArgument(Arguments.PACKAGE_NAME_OVERRIDES_PATH)));
+		this.addIfEnabled(context, Arguments.DISABLE_MOJMAP, () -> new MojmapNameProposer(context.getSingleArgument(Arguments.MOJMAP_PATH)));
 	}
 
 	@Override
