+++ src/main/java/org/quiltmc/enigma_plugin/proposal/UncheckedProposalService.java
@@ -2,6 +2,8 @@ package org.quiltmc.enigma_plugin.proposal;
 
 import org.quiltmc.enigma.api.service.EnigmaServiceContext;
 import org.quiltmc.enigma.api.service.NameProposalService;
+import org.quiltmc.enigma.api.translation.mapping.EntryMapping;
+import org.quiltmc.enigma.api.translation.representation.entry.Entry;
 import org.quiltmc.enigma_plugin.Arguments;
 import org.quiltmc.enigma_plugin.QuiltEnigmaPlugin;
 import org.quiltmc.enigma_plugin.index.JarIndexer;
@@ -9,7 +11,7 @@ import org.quiltmc.enigma_plugin.index.JarIndexer;
 public class UncheckedProposalService  extends NameProposerService {
 	public UncheckedProposalService(JarIndexer indexer, EnigmaServiceContext<NameProposalService> context) {
 		super();
-		this.addIfEnabled(context, Arguments.DISABLE_MOJMAP, MojmapPackageProposer::new);
+		this.addIfEnabled(context, Arguments.DISABLE_MOJMAP, () -> new MojmapPackageProposer(context.getSingleArgument(Arguments.PACKAGE_NAME_OVERRIDES_PATH)));
 	}
 
 	@Override
@@ -18,7 +20,7 @@ public class UncheckedProposalService  extends NameProposerService {
 	}
 
 	@Override
-	public boolean bypassValidation() {
-		return true;
+	public void validateProposedMapping(Entry<?> entry, EntryMapping mapping) {
+		// no-op
 	}
 }
