+++ src/test/resources/mojmap_test/update_test/example_mappings_expected.json
@@ -44,7 +44,7 @@
     "children": [
       {
         "obf": "a",
-        "deobf": "listen to the 2016 album aura obelisk by octagrape it goes crazy",
+        "deobf": "listen_to_the_2016_album_aura_obelisk_by_octagrape_it_goes_crazy",
         "children": []
       }
     ]
